+++ database/migrations/2021_04_14_140042_create_user_kyc_applications_table.php
@@ -43,6 +43,8 @@ class CreateUserKycApplicationsTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('user_kyc_applications');
+        if (Schema::hasTable('user_kyc_applications')) {
+            Schema::dropIfExists('user_kyc_applications');
+        }
     }
 }
