+++ database/migrations/2021_04_14_140226_create_cities_table.php
@@ -33,6 +33,8 @@ class CreateCitiesTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('cities');
+        if (Schema::hasTable('cities')) {
+            Schema::dropIfExists('cities');
+        }
     }
 }
